requirements.txt
aiohttp>=3.9.0
feedparser>=6.0.10
google-generativeai>=0.3.0
APScheduler>=3.10.4